refactor(TicketMetadata): extract to separate file
